+++ lib/formdata/index.ts
@@ -1,0 +1,8 @@
+'use strict';
+
+import {isFormData} from './is';
+
+export {
+    isFormData,
+    isFormData as is,
+};