+++ test/lib/formdata/index.test.ts
@@ -1,0 +1,15 @@
+'use strict';
+
+import {describe, it}   from 'node:test';
+import * as assert      from 'node:assert/strict';
+import * as LibFormData from '../../../lib/formdata';
+import is               from '../../../lib/formdata/is';
+
+describe('FormData - *', () => {
+    it('Should be a correct export', () => {
+        assert.deepEqual(LibFormData, {
+            isFormData: is,
+            is,
+        });
+    });
+});